requirements.txt
autopep8==1.6.0
bcrypt==3.2.2
cryptography==37.0.2
dnspython==2.2.1
fastapi==0.78.0
passlib==1.7.4
pydantic==1.9.1
pymongo==4.1.1
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
uvicorn==0.17.6
